move 'ACESlib' files to lib directory and change prefix for utility files to 'ACESutil'
The utilities directory contained CTL transforms identified as 'ACESlib' files. This is somewhat of a misnomer, as many of these files labeled 'ACESlib' are actually utility files designed for being run from the command line, e.g. Adjust_Exposure or the Lin_2_Log2 shaper functions. This commit creates a separate 'lib' directory, relocates the actual ACESlib files to there, and then renames utility files in the utilities directory with a more descriptive prefix of 'ACESutil'.

ACES documentation is also updated where appropriate to reflect the new prefix of 'ACESutil' for ACES transforms supplied with the ACES release.