make explicit the validation requirement for both minimum in/out and maximum in/out value pairs in Range node
Previously, the end of the description stated that "In both instances" values must be set such that maxOutValue = maxInValue × bitDepthScale. However, that is a bit of a mismatched statement because if only the minimum value pair is provided, then the rule is actually minOutValue = minInValue × bitDepthScale. We could have rewritten the "both instances" statement to say something more generic like OutValue = InValue × bitDepthScale, but the choice was made to put an explicit statement after each formula.